requirements.txt
slackclient==2.1.0
slackeventsapi==2.1.0
